+++ relay-server/src/utils/unreal.rs
@@ -17,6 +17,21 @@ use crate::envelope::{AttachmentType, ContentType, Envelope, Item, ItemType};
 /// Maximum number of unreal logs to parse for breadcrumbs.
 const MAX_NUM_UNREAL_LOGS: usize = 40;
 
+const SENTRY_PAYLOAD_KEY: &str = "__sentry";
+
+fn get_event_item(data: &[u8]) -> Result<Option<Item>, Unreal4Error> {
+    let mut context = Unreal4Context::parse(data)?;
+    let json = match context.game_data.remove(SENTRY_PAYLOAD_KEY) {
+        Some(json) if !json.is_empty() => json,
+        _ => return Ok(None),
+    };
+
+    log::trace!("adding event payload from unreal context");
+    let mut item = Item::new(ItemType::Event);
+    item.set_payload(ContentType::Json, json);
+    Ok(Some(item))
+}
+
 /// Expands Unreal 4 items inside an envelope.
 ///
 /// If the envelope does NOT contain an `UnrealReport` item, it doesn't do anything. If the envelope
@@ -32,6 +47,10 @@ pub fn expand_unreal_envelope(
     let payload = unreal_item.payload();
     let crash = Unreal4Crash::parse(&payload)?;
 
+    let mut has_event = envelope
+        .get_item_by(|item| item.ty() == ItemType::Event)
+        .is_some();
+
     for file in crash.files() {
         let (content_type, attachment_type) = match file.ty() {
             Unreal4FileType::Minidump => (ContentType::Minidump, AttachmentType::Minidump),
@@ -41,13 +60,20 @@ pub fn expand_unreal_envelope(
             Unreal4FileType::Log => (ContentType::Text, AttachmentType::UnrealLogs),
             Unreal4FileType::Config => (ContentType::OctetStream, AttachmentType::Attachment),
             Unreal4FileType::Context => (ContentType::Xml, AttachmentType::UnrealContext),
-            Unreal4FileType::Unknown => match file.name() {
+            _ => match file.name() {
                 self::ITEM_NAME_EVENT => (ContentType::MsgPack, AttachmentType::EventPayload),
                 self::ITEM_NAME_BREADCRUMBS1 => (ContentType::MsgPack, AttachmentType::Breadcrumbs),
                 self::ITEM_NAME_BREADCRUMBS2 => (ContentType::MsgPack, AttachmentType::Breadcrumbs),
                 _ => (ContentType::OctetStream, AttachmentType::Attachment),
             },
         };
+
+        if !has_event && attachment_type == AttachmentType::UnrealContext {
+            if let Some(event_item) = get_event_item(file.data())? {
+                envelope.add_item(event_item);
+                has_event = true;
+            }
+        }
 
         let mut item = Item::new(ItemType::Attachment);
         item.set_filename(file.name());
@@ -196,6 +222,20 @@ fn merge_unreal_context(event: &mut Event, context: Unreal4Context) {
 
     // modules not used just remove it from runtime props
     runtime_props.modules.take();
+
+    // promote all game data (except the special __sentry key) into a context.
+    if !context.game_data.is_empty() {
+        let game_context = contexts.get_or_insert_with("game", || Context::Other(Object::new()));
+        if let Context::Other(game_context) = game_context {
+            let filtered_keys = context
+                .game_data
+                .into_iter()
+                .filter(|(key, _)| key != SENTRY_PAYLOAD_KEY)
+                .map(|(key, value)| (key, Annotated::new(Value::String(value))));
+
+            game_context.extend(filtered_keys);
+        }
+    }
 
     if let Ok(Some(Value::Object(props))) = types::to_value(&runtime_props) {
         let unreal_context =